+++ src/js/lib/card-nicknames/mid.ts
@@ -1,0 +1,187 @@
+// Zendikar Rising (ZNR)
+export default [
+  {
+    // https://scryfall.com/card/mid/6/uzumaki-deluxe?utm_source=api
+    setCode: "mid",
+    collectorNumber: "6",
+    realName: ["Borrowed Time"],
+    nickname: ["Uzumaki Deluxe"],
+    source: "Scryfall Preview Name",
+  },
+  {
+    // https://scryfall.com/card/mid/25/indomitable-cathar?utm_source=api
+    setCode: "mid",
+    collectorNumber: "25",
+    realName: ["Intrepid Adversary"],
+    nickname: ["Indomitable Cathar"],
+    source: "Scryfall Preview Name",
+  },
+  {
+    // https://scryfall.com/card/mid/38/night-at-the-roxbury?utm_source=api
+    setCode: "mid",
+    collectorNumber: "38",
+    realName: ["Sunset Revelry"],
+    nickname: ["Night at the Roxbury"],
+    source: "Scryfall Preview Name",
+  },
+  {
+    // https://scryfall.com/card/mid/41/odric-is-tired-of-your-shit?utm_source=api
+    setCode: "mid",
+    collectorNumber: "41",
+    realName: ["Vanquish the Horde"],
+    nickname: ["Odric is Tired of Your Shit"],
+    source: "Scryfall Preview Name",
+  },
+  {
+    // https://scryfall.com/card/mid/45/thief-of-the-coast-angery-ghost?utm_source=api
+    setCode: "mid",
+    collectorNumber: "45",
+    realName: ["Covetous Castaway", "Ghostly Castigator"],
+    nickname: ["Thief of the Coast", "Angery Ghost"],
+    source: "Scryfall Preview Name",
+  },
+  {
+    // https://scryfall.com/card/mid/58/pickle-jar-zombie?utm_source=api
+    setCode: "mid",
+    collectorNumber: "58",
+    realName: ["Larder Zombie"],
+    nickname: ["Pickle-Jar Zombie"],
+    source: "Scryfall Preview Name",
+  },
+  {
+    // https://scryfall.com/card/mid/71/geri-the-cleaner-als-toy-barn?utm_source=api
+    setCode: "mid",
+    collectorNumber: "71",
+    realName: ["Poppet Stitcher", "Poppet Factory"],
+    nickname: ["Geri the Cleaner", "Al's Toy Barn"],
+    source: "Scryfall Preview Name",
+  },
+  {
+    // https://scryfall.com/card/mid/75/innistragram-influencer?utm_source=api
+    setCode: "mid",
+    collectorNumber: "75",
+    realName: ["Skaab Wrangler"],
+    nickname: ["Innistragram Influencer"],
+    source: "Scryfall Preview Name",
+  },
+  {
+    // https://scryfall.com/card/mid/85/doomblade-scoundrel-doomclaw-marauder?utm_source=api
+    setCode: "mid",
+    collectorNumber: "85",
+    realName: ["Baneblade Scoundrel", "Baneclaw Marauder"],
+    nickname: ["Doomblade Scoundrel", "Doomclaw Marauder"],
+    source: "Scryfall Preview Name",
+  },
+  {
+    // https://scryfall.com/card/mid/89/exterminate-the-lineage?utm_source=api
+    setCode: "mid",
+    collectorNumber: "89",
+    realName: ["Bloodline Culling"],
+    nickname: ["Exterminate the Lineage"],
+    source: "Scryfall Preview Name",
+  },
+  {
+    // https://scryfall.com/card/mid/111/+2-griselbrand-skull?utm_source=api
+    setCode: "mid",
+    collectorNumber: "111",
+    realName: ["Mask of Griselbrand"],
+    nickname: ["+2 Griselbrand Skull"],
+    source: "Scryfall Preview Name",
+  },
+  {
+    // https://scryfall.com/card/mid/121/jump-scare-zombies?utm_source=api
+    setCode: "mid",
+    collectorNumber: "121",
+    realName: ["Jump Scare Zombies"],
+    nickname: ["Siege Zombie"],
+    source: "Scryfall Preview Name",
+  },
+  {
+    // https://scryfall.com/card/mid/137/terrible-housemate?utm_source=api
+    setCode: "mid",
+    collectorNumber: "137",
+    realName: ["Falkenrath Pit Fighter"],
+    nickname: ["Terrible Housemate"],
+    source: "Scryfall Preview Name",
+  },
+  {
+    // https://scryfall.com/card/mid/155/%23cleantok-dragon?utm_source=api
+    setCode: "mid",
+    collectorNumber: "155",
+    realName: ["Purifying Dragon"],
+    nickname: ["#cleanTok Dragon"],
+    source: "Scryfall Preview Name",
+  },
+  {
+    // https://scryfall.com/card/mid/173/mommy-spider?utm_source=api
+    setCode: "mid",
+    collectorNumber: "173",
+    realName: ["Brood Weaver"],
+    nickname: ["Mommy Spider"],
+    source: "Scryfall Preview Name",
+  },
+  {
+    // https://scryfall.com/card/mid/177/tarmogellyf?utm_source=api
+    setCode: "mid",
+    collectorNumber: "177",
+    realName: ["Consuming Blob"],
+    nickname: ["Tarmogellyf"],
+    source: "Scryfall Preview Name",
+  },
+  {
+    // https://scryfall.com/card/mid/178/aerial-stunt-doubles?utm_source=api
+    setCode: "mid",
+    collectorNumber: "178",
+    realName: ["Contortionist Troupe"],
+    nickname: ["Aerial Stunt Doubles"],
+    source: "Scryfall Preview Name",
+  },
+  {
+    // https://scryfall.com/card/mid/183/wren-and-feeling-it?utm_source=api
+    setCode: "mid",
+    collectorNumber: "183",
+    realName: ["Dryad's Revival"],
+    nickname: ["Wren and Feeling It"],
+    source: "Scryfall Preview Name",
+  },
+  {
+    // https://scryfall.com/card/mid/196/do-you-want-ants%3F-because%E2%80%94?utm_source=api
+    setCode: "mid",
+    collectorNumber: "196",
+    realName: ["Rise of the Ants"],
+    nickname: ["Do You Want Ants? Because—"],
+    source: "Scryfall Preview Name",
+  },
+  {
+    // https://scryfall.com/card/mid/216/fen-triplets?utm_source=api
+    setCode: "mid",
+    collectorNumber: "216",
+    realName: ["Dawnhart Wardens"],
+    nickname: ["Fen Triplets"],
+    source: "Scryfall Preview Name",
+  },
+  {
+    // https://scryfall.com/card/mid/226/that-thing-from-annihilation?utm_source=api
+    setCode: "mid",
+    collectorNumber: "226",
+    realName: ["Grizzly Ghoul"],
+    nickname: ["That Thing from Annihilation"],
+    source: "Scryfall Preview Name",
+  },
+  {
+    // https://scryfall.com/card/mid/238/angry-maro?utm_source=api
+    setCode: "mid",
+    collectorNumber: "238",
+    realName: ["Rootcoil Creeper"],
+    nickname: ["Angry Maro"],
+    source: "Scryfall Preview Name",
+  },
+  {
+    // https://scryfall.com/card/mid/247/spoopy-moontime?utm_source=api
+    setCode: "mid",
+    collectorNumber: "247",
+    realName: ["Unnatural Moonrise"],
+    nickname: ["Spoopy Moontime"],
+    source: "Scryfall Preview Name",
+  },
+];